Remove squashed migration and rely on `run_before`
The initial migration must continue to rely on the historical
PageRevision model. Changing it to the generic revision model may cause
errors with existing databases that have migrations applied in a certain
order.

New databases won't be able to migrate because they have no history and
try to migrate the initial migration using the PageRevision model _after_
it was removed by Wagtail's migrations. To combat this, we add `run_before`
to ensure we run the initial migration _before_ the PageRevision model is
removed by Wagtail's migrations.

I feel like this is a better solution then squashing the migration
history.