Revert "Remove squashed migration and rely on `run_before`"
This reverts commit 4e0e127b010e634a1050f7e1cc480d49ebca8f06.

The run_before approach does not work properly if ab_testing is
installed in a project that is already on Wagtail 4 and has the
PageRevision to Revision migration applied. In this case, is not
possible to make our initial migration run before Wagtail's migration.

Thanks to Saptak Sengupta for pointing this out.